+++ resources/js/plugins/axios.js
@@ -8,7 +8,7 @@ import i18n from '~/plugins/i18n'
 axios.interceptors.request.use(request => {
   const token = store.getters['auth/token']
   if (token) {
-    request.headers.common['Authorization'] = `Bearer ${token}`
+    request.headers.common.Authorization = `Bearer ${token}`
   }
 
   const locale = store.getters['lang/locale']
@@ -27,7 +27,7 @@ axios.interceptors.response.use(response => response, error => {
 
   if (status >= 500) {
     Swal.fire({
-      type: 'error',
+      icon: 'error',
       title: i18n.t('error_alert_title'),
       text: i18n.t('error_alert_text'),
       reverseButtons: true,
@@ -38,7 +38,7 @@ axios.interceptors.response.use(response => response, error => {
 
   if (status === 401 && store.getters['auth/check']) {
     Swal.fire({
-      type: 'warning',
+      icon: 'warning',
       title: i18n.t('token_expired_alert_title'),
       text: i18n.t('token_expired_alert_text'),
       reverseButtons: true,