+++ routes/spa.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\SpaController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -13,4 +14,4 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('{path}', 'SpaController')->where('path', '(.*)');
+Route::get('{path}', SpaController::class)->where('path', '(.*)');