+++ webpack.mix.js
@@ -31,12 +31,14 @@ mix.webpackConfig({
   },
   output: {
     chunkFilename: 'dist/js/[chunkhash].js',
-    path: mix.config.hmr ? '/' : path.resolve(__dirname, './public/build')
+    path: mix.config.hmr
+      ? '/'
+      : path.resolve(__dirname, mix.inProduction() ? './public/build' : './public')
   }
 })
 
 mix.then(() => {
-  if (!mix.config.hmr) {
+  if (mix.inProduction()) {
     process.nextTick(() => publishAseets())
   }
 })
@@ -44,10 +46,7 @@ mix.then(() => {
 function publishAseets () {
   const publicDir = path.resolve(__dirname, './public')
 
-  if (mix.inProduction()) {
-    fs.removeSync(path.join(publicDir, 'dist'))
-  }
-
+  fs.removeSync(path.join(publicDir, 'dist'))
   fs.copySync(path.join(publicDir, 'build', 'dist'), path.join(publicDir, 'dist'))
   fs.removeSync(path.join(publicDir, 'build'))
 }